Implement pareto rule for more restrictive SD evaluation
A simple pareto rule can be that animals spending > 80 % of time (30 minute windows) in < 20 % of the ROI are likely to be asleep or at least not moving at all. This assumes most of the ROI (>75% is accessible to the fly)
This avoids labelling a fly as awake when

* it is moving with the tube as it is deprived, but it does not move by itself

* it is not moving but changes in the background during SD make it difficult for the tracker to detect the animal correctly (spiky artifact)

Other features of the spiky artifact could be used to label a time series as problematic. However, many of them are expensive to compute or difficult to program. This pareto rule is easy to program, fast to run and also easy to explain. And it seems to give good results. Also for this application we dont mind being too restrictive. Discarding too many flies is better than picking a "SDd" fly which is not really SDd!

Also start implementing loading of QC table